refactor network utils
